+++ packages/adblocker-puppeteer/rollup.config.ts
@@ -7,6 +7,7 @@
  */
 
 import resolve from 'rollup-plugin-node-resolve';
+import sourcemaps from 'rollup-plugin-sourcemaps';
 
 export default {
   // CommonJS
@@ -17,5 +18,8 @@ export default {
     format: 'cjs',
     sourcemap: true,
   },
-  plugins: [resolve()],
+  plugins: [
+    resolve(),
+    sourcemaps(),
+  ],
 };